# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing buildings and passes.",
        "title": "Service-propusk API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "172.17.17.145:8000",
    "basePath": "/",
    "paths": {
...
        "/passes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список пропусков с возможностью фильтрации по статусу и датам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Получить список пропусков",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Статус пропуска для фильтрации",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начальная дата для фильтрации (формат: DD.MM.YYYY)",
                        "name": "begin_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата для фильтрации (формат: DD.MM.YYYY)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/passhandler.PassShortResp"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/passes/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает полную информацию о конкретном пропуске",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Получить пропуск по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пропуска (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/passhandler.PassResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию о существующем пропуске",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Обновить пропуск",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пропуска (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные пропуска",
                        "name": "pass",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passhandler.EditPassRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет пропуск из системы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Удалить пропуск",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пропуска (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "passhandler.Building": {
            "description": "Информация о здании в пропуске",
            "type": "object",
            "properties": {
                "description": {
                    "description": "Описание здания",
                    "type": "string",
                    "example": "Многоэтажное офисное здание"
                },
                "id": {
                    "description": "ID здания",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "imgUrl": {
                    "description": "URL изображения здания",
                    "type": "string",
                    "example": "http://example.com/image.jpg"
                },
                "name": {
                    "description": "Название здания",
                    "type": "string",
                    "example": "Офисное здание"
                }
            }
        },
        "passhandler.EditPassRequest": {
            "description": "Запрос на обновление информации о пропуске",
            "type": "object",
            "properties": {
                "date_visit": {
                    "description": "Дата посещения",
                    "type": "string",
                    "example": "2024-03-10T12:00:00Z"
                },
                "visitor": {
                    "description": "Имя посетителя",
                    "type": "string",
                    "example": "Иван Иванов"
                }
            }
        },
        "passhandler.ErrorResponse": {
            "description": "Структура ответа при возникновении ошибки",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Текст ошибки",
                    "type": "string",
                    "example": "Описание ошибки"
                }
            }
        },
        "passhandler.PassItem": {
            "description": "Элемент пропуска (здание с комментарием)",
            "type": "object",
            "properties": {
                "building": {
                    "description": "Информация о здании",
                    "allOf": [
                        {
                            "$ref": "#/definitions/passhandler.Building"
                        }
                    ]
                },
                "comment": {
                    "description": "Комментарий к зданию",
                    "type": "string",
                    "example": "Комментарий к зданию"
                }
            }
        },
        "passhandler.PassResp": {
            "description": "Полная информация о пропуске",
            "type": "object",
            "properties": {
                "dateVisit": {
                    "description": "Дата посещения",
                    "type": "string",
                    "example": "2024-03-10T12:00:00Z"
                },
                "id": {
                    "description": "ID пропуска",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "items": {
                    "description": "Список зданий в пропуске",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/passhandler.PassItem"
                    }
                },
                "visitorName": {
                    "description": "Имя посетителя",
                    "type": "string",
                    "example": "Иван Иванов"
                }
            }
        },
        "passhandler.PassShortResp": {
            "description": "Краткая информация о пропуске для списка",
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID пропуска",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "moderator": {
                    "description": "Информация о модераторе (если есть)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/passhandler.User"
                        }
                    ]
                },
                "status": {
                    "description": "Статус пропуска",
                    "type": "integer",
                    "example": 1
                },
                "user": {
                    "description": "Информация о создателе пропуска",
                    "allOf": [
                        {
                            "$ref": "#/definitions/passhandler.User"
                        }
                    ]
                }
            }
        },
        "passhandler.User": {
            "description": "Информация о пользователе",
            "type": "object",
            "properties": {
                "login": {
                    "description": "Логин пользователя",
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "cookieAuth": {
            "description": "Enter your JWT token value",
            "type": "apiKey",
            "name": "X-Access-Token",
            "in": "header"
        }
    }
}